farpatch: add `#include "platform.h"` where needed
The general header files no longer include `platform.h` as part of the
general include set. This results in lots of errors when building with
the latest blackmagic project.

Additionally, `RAM_ATTR` now requires including `esp_attr.h`.

Fix both sets of includes in a variety of files in order to fix the
build.

Signed-off-by: Sean Cross <sean@xobs.io>